bgpd: deferral timer enlarge encoding style
the bgp deferral timer is set to 16 bit size. as deferral values can
range from 0 to 720000, this is far over the 16 bit size. Theorically,
this can be a problem for bgp deferral values greater than 65 seconds.
Increase the size of deferral timer from 16 to 32 bit size.

+++ bgpd/bgpd.h
@@ -220,7 +220,7 @@ struct bgp
 #define MAX_EOR_UPDATE_DELAY 3600
   u_int16_t v_update_delay;
 #define MAX_BGP_SELECTION_DEFERRAL 720000
-  u_int16_t v_selection_deferral;
+  u_int32_t v_selection_deferral;
 
   QZC_NODE
 };
